feat(components): move Modal.js and Modal.css to Modal directory
