+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import Modal from "./components/Modal";
+import Modal from "./components/Modal/Modal";
 import AddSkillForm from "./views/Skills/AddSkillForm";
 import "./App.css";
